feat-Update mnist cpp trainer
* Change input mnist label data type from int64 -> float32
* Use one-hot encoding for mnist label data, details can be found here (https://www.quora.com/What-does-the-one_hot-True-parameter-on-the-MNIST-tensorflow-for-beginners-example-mean)

+++ tensorflow/examples/train_mnist/mnistReader.h
@@ -19,11 +19,11 @@ public:
 
   // Training Batch
   void getTrainingBatch(int beginIdx, int batchSize,
-    vector<float>* batchImgVec, vector<long int>* batchLabelVec);
+    vector<float>* batchImgVec, vector<float>* batchLabelVec);
 
   // Testing Batch
   void getTestingBatch(int beginIdx, int batchSize,
-    vector<float>* batchImgVec, vector<long int>* batchLabelVec);
+    vector<float>* batchImgVec, vector<float>* batchLabelVec);
 
   // getTrainingDataSize
   int getTrainingDataSize();
